Last two problems left

+++ Practice Floating Point Numbers/Program.cs
@@ -6,7 +6,7 @@ namespace Practice_Floating_Point_Numbers
     {
         static void Main(string[] args)
         {
-            double num1 = 3.141592653589793238;
+            decimal num1 = 3.141592653589793238m;
             double num2 = 1.60217657;
             decimal num3 = 7.8184261974584555216535342341m;
